Draft adding ANA flowfiles to iceye collection creation
A class AnaFlowProcessor was created in bench.py that is meant to be
able to be generically used for collections when an ANA flow file is
needed and an event datetime or date range is available. The class takes
the nwm_flows gpkg with nwm features and connects to the NWC's google
cloud bucket to find AnA forecasts associated with a given date range in
either CONUS, Hawaii, or Alaska. Then for a datetime the closest
matching AnA netcdf forecast file is used to create a flowfile for a
given STAC item's observations using the STAC item's geometry to
restrict which reach id's are assigned discharges. If a date range is
provided then the peak observed discharge within the AOI is used to
select a forecast file to extract discharges from.

This class was used in the iceye collection creation scripts to create
ANA derived flowfiles for the historical events in the iceye collection.
This involved updating all three iceye scripts to allow for the creation
and uploading of the flowfiles and then indexing the newly created
flowfile as an asset in the iceye collection stac items