Iceye flows (#4)
* Add iceye collection generation scripts, add top level README.md, and iceye specific README.md files

* Modifications to succesffully create iceye collection. Update README.md, add unzip.py used for extracting original .zip files.

* Update bench.py, schemas/iceye/*, and iceye collection creation.

Improve RasterUtils.create_preview in bench.py to use rasertio's decimated read, for faster operations.
Use this method for iceye depth thumbnail creation (using raster), instead of using extents (vectors & matplotlib) to
have consistency across collection creation.
Update schemas appropriately.

* Reformat recreate_collection.sh to take ITEM_DIR and COLLECTION_JSON as
parameters instead of hard-coding values.

* Update README's and delete old scripts

Edited main README to add aditional context and remove some repetition.

Also performed a minor edit to iceye readme.

Removed 3 old scripts that were created before the requirements for the
benchmark STAC functionality were fully solidified.

* Fix duplicate link creation in catalog.json

Added deduplication for child links in the
download_catalog_and_collections in bench.py. This addresses the
duplicate child link creation that was being observed on collection
update even when using pystac's remove_child method

* Simplify requirements.txt

Only stipulate direct dependencies imported by python scripts in repo.
Add versioning constraints

* Draft adding ANA flowfiles to iceye collection creation

A class AnaFlowProcessor was created in bench.py that is meant to be
able to be generically used for collections when an ANA flow file is
needed and an event datetime or date range is available. The class takes
the nwm_flows gpkg with nwm features and connects to the NWC's google
cloud bucket to find AnA forecasts associated with a given date range in
either CONUS, Hawaii, or Alaska. Then for a datetime the closest
matching AnA netcdf forecast file is used to create a flowfile for a
given STAC item's observations using the STAC item's geometry to
restrict which reach id's are assigned discharges. If a date range is
provided then the peak observed discharge within the AOI is used to
select a forecast file to extract discharges from.

This class was used in the iceye collection creation scripts to create
ANA derived flowfiles for the historical events in the iceye collection.
This involved updating all three iceye scripts to allow for the creation
and uploading of the flowfiles and then indexing the newly created
flowfile as an asset in the iceye collection stac items

* Split bench.py into utils.py and flows.py

bench.py was growing too long so split into two files

---------

Co-authored-by: robgpita-noaa <rob.g.pita@noaa.gov>
Co-authored-by: robgpita <rob.gonzalez-pita@ertcorp.com>